refactor(ui): move port inspection out of model.go
Port scanning and the external-conflict dialog form a self-contained theme:
they read the port package and own one modal, and touch dashboard state only
through the focused service.

Declarations move unchanged.
